app/agents/core_agents/business_agent.py
import logging
from typing import Any, Dict, List, Tuple
from app.agents.core_agents.agent_types import Agent
from app.models.agent.belief import Belief
from app.models.agent.desire import Desire
from app.models.agent.intention import Intention
from app.models.agent.goal import Goal
from app.models.rules.rules_engine import RulesEngine
from app.models.knowledge.knowledge_graph import KnowledgeGraph
from app.models.knowledge.reasoning.temporal_reasoning import TemporalReasoner
from app.models.knowledge.conflict_resolution import ConflictResolver
from app.models.knowledge.explanation_generator import ExplanationGenerator
from app.tools.reasoning_engine import ReasoningEngine
from datetime import datetime
from openai import OpenAI
from app.models.business.business_goal import BusinessGoalCreate, BusinessGoalResponse
from app.db.arango_db_client import ArangoDBClient
from uuid import uuid4

logger = logging.getLogger(__name__)

class BusinessAgent(Agent):
    business_id: str = Field(..., description="The business ID associated with the agent")
    business_type: str = Field(..., description="The type of business this agent represents")
    financial_data: Dict[str, Any] = Field(default_factory=dict, description="Financial data of the business")
    rules_engine: RulesEngine = Field(..., description="The rules engine used for decision-making")
    knowledge_graph: KnowledgeGraph = Field(..., description="The knowledge graph used for enriching context")
    temporal_reasoner: TemporalReasoner = Field(..., description="The temporal reasoner used for evaluating rules over time")
    conflict_resolver: ConflictResolver = Field(..., description="The conflict resolver used for resolving conflicts in rule outcomes")
    explanation_generator: ExplanationGenerator = Field(..., description="The explanation generator used for generating explanations for decisions")
    reasoning_engine: ReasoningEngine = Field(..., description="The reasoning engine used for simulating scenarios and generating action plans")

    def __init__(self, 
                 rules_engine: RulesEngine, 
                 knowledge_graph: KnowledgeGraph,
                 temporal_reasoner: TemporalReasoner,
                 conflict_resolver: ConflictResolver,
                 explanation_generator: ExplanationGenerator,
                 reasoning_engine: ReasoningEngine,
                 **data):
        super().__init__(**data)
        self.rules_engine = rules_engine
        self.knowledge_graph = knowledge_graph
        self.temporal_reasoner = temporal_reasoner
        self.conflict_resolver = conflict_resolver
        self.explanation_generator = explanation_generator
        self.reasoning_engine = reasoning_engine
        self._init_business_beliefs()
        self._init_business_desires()

    def _init_business_beliefs(self):
        self.add_belief(Belief(id="business_id", content={"id": self.business_id}, description=f"Business ID: {self.business_id}", certainty=1.0))
        self.add_belief(Belief(id="business_type", content={"type": self.business_type}, description=f"Business Type: {self.business_type}", certainty=1.0))

    def _init_business_desires(self):
        self.add_desire(Desire(id="increase_profit", description="Increase business profit", priority=0.9))
        self.add_desire(Desire(id="expand_market", description="Expand market share", priority=0.8))

    async def make_decision(self, context: Dict[str, Any]) -> Tuple[List[Tuple[str, Any]], str]:
        # Enrich context with knowledge graph
        enriched_context = self.knowledge_graph.enrich_context(context)

        # Evaluate rules
        results = self.rules_engine.evaluate_rules_with_control(enriched_context)

        # Resolve conflicts if any
        resolved_results = self.conflict_resolver.resolve(results)

        # Generate explanation
        explanation = self.explanation_generator.explain_decision(self.rules_engine, enriched_context, resolved_results)

        return resolved_results, explanation

    async def analyze_rule_outcomes(self, rule_name: str) -> str:
        truth_table = self.rules_engine.generate_truth_table(rule_name)
        analysis = await self.rules_engine.reason_about_truth_table(rule_name)
        return analysis

    async def update_business_rules(self, new_business_description: str) -> str:
        new_rules = await self.rules_engine.generate_rules_from_description(new_business_description)
        return f"Updated rules based on new description: {new_rules}"

    async def get_rule_insights(self, query: str) -> str:
        insights = await self.rules_engine.reason_about_rules(query)
        return insights

    async def evaluate_rules_over_time(self, params: Dict[str, Any], start_time: datetime, end_time: datetime) -> List[Tuple[datetime, List[Tuple[str, Any]]]]:
        return self.temporal_reasoner.evaluate_rules_in_timeframe(self.rules_engine, params, (start_time, end_time))

    async def reason_about_decision(self, context: Dict[str, Any], decision: List[Tuple[str, Any]]) -> str:
        reasoning_prompt = f"Given the context {context} and the decision {decision}, provide a detailed reasoning about why this decision was made and its potential implications."
        reasoning = await self.reasoning_engine.reason(reasoning_prompt)
        return reasoning

    async def simulate_scenario(self, scenario: Dict[str, Any]) -> str:
        # Use the reasoning engine to simulate a business scenario
        simulation_prompt = f"Given the following business scenario: {scenario}, simulate the potential outcomes and their implications for the business."
        simulation_result = await self.reasoning_engine.reason(simulation_prompt)
        return simulation_result

    async def generate_action_plan(self, goal: str, context: Dict[str, Any]) -> List[str]:
        # Use the reasoning engine to generate an action plan
        plan_prompt = f"Given the goal '{goal}' and the current business context {context}, generate a step-by-step action plan to achieve this goal."
        action_plan = await self.reasoning_engine.reason(plan_prompt)
        return action_plan.split('\n')  # Assuming the plan is returned as a newline-separated list

    async def assess_risk(self, decision: List[Tuple[str, Any]], context: Dict[str, Any]) -> str:
        # Use the reasoning engine to assess potential risks
        risk_prompt = f"Given the decision {decision} in the context {context}, assess the potential risks and their likelihood."
        risk_assessment = await self.reasoning_engine.reason(risk_prompt)
        return risk_assessment

    def execute_action(self, action: str, params: Dict[str, Any]) -> Any:
        # This method would contain the logic to actually execute business actions
        # For now, we'll just print the action and return a dummy result
        print(f"Executing action: {action} with parameters: {params}")
        return f"Action {action} executed successfully"

    async def reason(self, context: Dict[str, Any]) -> Dict[str, Any]:
        reasoning_result = await super().reason(context)
        
        # Business-specific reasoning logic
        if "financial_data" in context:
            self._update_financial_beliefs(context["financial_data"])
        
        return reasoning_result

    def _update_financial_beliefs(self, financial_data: Dict[str, Any]):
        for key, value in financial_data.items():
            self.add_belief(Belief(id=f"financial_{key}", content={key: value}, description=f"Financial data: {key}", certainty=0.9))

    async def act(self) -> None:
        """Perform the agent's main action cycle: reason, deliberate, and execute business actions."""
        try:
            context = self.get_current_state()
            await self.reason(context)
            self.deliberate()
            await self._execute_business_actions()
        except Exception as e:
            logger.error(f"Error during business agent action cycle: {str(e)}")

    def deliberate(self) -> None:
        """
        Deliberate on current beliefs and desires to form business-related intentions.
        """
        for desire in self.desires:
            if desire.id == "increase_profit" and self._should_focus_on_profit():
                self.add_intention(Intention(id="implement_cost_cutting", goal=Goal(id="reduce_costs", description="Reduce operational costs"), plan=None))
            elif desire.id == "expand_market" and self._should_focus_on_expansion():
                self.add_intention(Intention(id="launch_marketing_campaign", goal=Goal(id="increase_market_share", description="Increase market share"), plan=None))

    def _should_focus_on_profit(self) -> bool:
        # Logic to determine if the agent should focus on increasing profit
        return any(belief.content.get("profit_margin", 0) < 0.2 for belief in self.beliefs if "profit_margin" in belief.content)

    def _should_focus_on_expansion(self) -> bool:
        # Logic to determine if the agent should focus on market expansion
        return any(belief.content.get("market_share", 0) < 0.3 for belief in self.beliefs if "market_share" in belief.content)

    async def _execute_business_actions(self):
        for intention in self.intentions:
            if intention.id == "implement_cost_cutting":
                await self._implement_cost_cutting()
            elif intention.id == "launch_marketing_campaign":
                await self._launch_marketing_campaign()

    async def _implement_cost_cutting(self):
        # Implement cost-cutting measures
        print("Implementing cost-cutting measures")
        # In a real system, this would involve more complex logic and possibly interaction with other services

    async def _launch_marketing_campaign(self):
        # Launch a marketing campaign
        print("Launching marketing campaign")
        # In a real system, this would involve more complex logic and possibly interaction with other services

    def create_business_goal(self, goal_data: BusinessGoalCreate) -> BusinessGoalResponse:
        # Generate a unique ID for the new goal
        goal_id = str(uuid4())

        # Use OpenAI to validate and enhance the goal description
        completion = client.beta.chat.completions.parse(
            model="gpt-4o-2024-08-06",
            messages=[
                {"role": "system", "content": "You are an AI assistant helping to create and validate business goals."},
                {"role": "user", "content": f"Please validate and enhance the following business goal: {goal_data.dict()}"}
            ],
            response_format=BusinessGoalResponse
        )

        if hasattr(completion.choices[0].message, 'refusal'):
            print(completion.choices[0].message.refusal)
            return BusinessGoalResponse(goal=None, message="Failed to create business goal due to AI refusal.")

        result = completion.choices[0].message.parsed

        # Create the new BusinessGoal object
        new_goal = BusinessGoal(
            id=goal_id,
            **goal_data.dict(),
            child_goal_ids=[],
            kpis=[],
            metadata={}
        )

        # Save the new goal to the knowledge graph
        self.knowledge_graph.add_business_goal(new_goal)

        return BusinessGoalResponse(goal=new_goal, message="Business goal created successfully.")

    def get_business_goal(self, goal_id: str) -> BusinessGoal:
        goal_data = self.knowledge_graph.get_business_goal(goal_id)
        return BusinessGoal(**goal_data)

    def update_business_goal(self, goal_id: str, updated_data: Dict[str, Any]) -> BusinessGoalResponse:
        self.knowledge_graph.update_business_goal(goal_id, updated_data)
        updated_goal = self.get_business_goal(goal_id)
        return BusinessGoalResponse(goal=updated_goal, message="Business goal updated successfully.")

    def delete_business_goal(self, goal_id: str) -> str:
        self.knowledge_graph.delete_business_goal(goal_id)
        return f"Business goal {goal_id} deleted successfully."

    def get_all_business_goals(self) -> List[BusinessGoal]:
        goals_data = self.knowledge_graph.get_all_business_goals()
        return [BusinessGoal(**goal_data) for goal_data in goals_data]

    def get_child_goals(self, parent_goal_id: str) -> List[BusinessGoal]:
        child_goals_data = self.knowledge_graph.get_child_goals(parent_goal_id)
        return [BusinessGoal(**goal_data) for goal_data in child_goals_data]

# Additional functions for updating, deleting, and retrieving business goals can be implemented similarly